Significant refactor and improvement to seperation of concerns
