+++ atlas-extension/project-atlas/src/app/contexts/Camera.ts
@@ -1,5 +1,7 @@
 import * as React from "react";
 import { Camera } from "../models/camera";
 
-const CameraContext = React.createContext<Camera | null>(null);
-export default CameraContext;
+export type CameraManager = [Camera, (newCamera: Partial<Camera>) => void];
+
+const CameraManagerContext = React.createContext<CameraManager | null>(null);
+export default CameraManagerContext;
