+++ atlas-extension/project-atlas/src/app/contexts/Project.ts
@@ -1,0 +1,7 @@
+import * as React from "react";
+import { Project } from "../models/project";
+
+export type ProjectManager = [Project | null, (newProject: Project | null) => void];
+
+const ProjectManagerContext = React.createContext<ProjectManager | null>(null);
+export default ProjectManagerContext;
