+++ atlas-extension/project-atlas/src/app/models/mouse.ts
@@ -1,0 +1,5 @@
+export enum Mouse {
+    SingleClick = 1,
+    DoubleClick = 2,
+    TripleClick = 3,
+}
